Added worst case transmission time and fixed opt
I added worst case transmission time for a data flow to the model. I
have it commented it out though because for right now we are getting out
of memory errors with Java. I also removed the optimization goals for
maximizing the WCET while minimizing the TCs. When we optimize the TCs
we find the solution in a reasonable amount of time.